Start matching operators to the same structure as units
